# docs/api/openapi-schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kàff GUI API",
        "version": "1.0.0",
        "description": "API REST de gestion de volière colombophile.\n\n## Authentification\n\nL'API utilise JWT (JSON Web Tokens) avec authentification 2FA (TOTP).\n\n### Flux d'authentification\n\n1. **Inscription** : `POST /api/v1/auth/register/` - Retourne un QR code TOTP\n2. **Confirmation TOTP** : `POST /api/v1/auth/2fa/confirm/` - Active le compte\n3. **Connexion** : `POST /api/v1/auth/login/` - Retourne access + refresh tokens\n4. **Utilisation** : Ajouter `Authorization: Bearer <access_token>` dans les headers\n5. **Refresh** : `POST /api/v1/auth/token/refresh/` - Obtenir un nouveau access token\n\n### Sécurité\n\n- Tokens access : 15 minutes de validité\n- Tokens refresh : 7 jours de validité avec rotation\n- Rate limiting : 300 req/min pour utilisateurs authentifiés\n- Protection brute-force : 5 tentatives max, cooldown 5 min\n\nVoir le cahier des charges et les user stories pour le détail métier.",
        "contact": {
            "name": "Serigne Mourtalla Syll",
            "email": "serignemourtallasyll86@gmail.com"
        },
        "license": {
            "name": "Proprietary"
        }
    },
    "paths": {
...
        "/api/v1/cages/{id}/affecter-couple/": {
            "post": {
                "operationId": "cages_affecter_couple_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Cage.",
                        "required": true
                    }
                ],
                "tags": [
                    "cages"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AffecterCoupleInputRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AffecterCoupleInputRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AffecterCoupleInputRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/cages/{id}/affecter-pigeon/": {
            "post": {
                "operationId": "cages_affecter_pigeon_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Cage.",
                        "required": true
                    }
                ],
                "tags": [
                    "cages"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AffecterPigeonInputRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AffecterPigeonInputRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AffecterPigeonInputRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/cages/{id}/liberer/": {
            "post": {
                "operationId": "cages_liberer_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Cage.",
                        "required": true
                    }
                ],
                "tags": [
                    "cages"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CageRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CageRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/cages/volet/": {
            "get": {
                "operationId": "cages_volet_list",
                "description": "Renvoie l'état de toutes les cages avec leur occupant courant. Code couleur : vert=libre, rouge=pigeon, orange=couple. Optimisé en 1 requête + 1 prefetch (zéro N+1).",
                "parameters": [
                    {
                        "in": "query",
                        "name": "numero",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "numero__icontains",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ordering",
                        "required": false,
                        "in": "query",
                        "description": "Which field to use when ordering the results.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "A page number within the paginated result set.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page_size",
                        "required": false,
                        "in": "query",
                        "description": "Number of results to return per page.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "required": false,
                        "in": "query",
                        "description": "A search term.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "statut_occupation",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "LIBRE",
                                "OCCUPE_COUPLE",
                                "OCCUPE_PIGEON"
                            ]
                        },
                        "description": "* `LIBRE` - Libre\n* `OCCUPE_PIGEON` - Occupée par un pigeon\n* `OCCUPE_COUPLE` - Occupée par un couple"
                    },
                    {
                        "in": "query",
                        "name": "statut_occupation__in",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "description": "Les valeurs multiples doivent être séparées par des virgules.",
                        "explode": false,
                        "style": "form"
                    }
                ],
                "tags": [
                    "cages"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedVolereCageList"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AffecterCoupleInputRequest": {
                "type": "object",
                "properties": {
                    "couple_id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "couple_id"
                ]
            },
            "AffecterPigeonInputRequest": {
                "type": "object",
                "properties": {
                    "pigeon_id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "pigeon_id"
                ]
            },
...
            "Cage": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "title": "Identifiant"
                    },
                    "numero": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "nom": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "superficie": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$",
                        "nullable": true
                    },
                    "description": {
                        "type": "string"
                    },
                    "statut_occupation": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StatutOccupationEnum"
                            }
                        ],
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "Créé le"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "Mis à jour le"
                    }
                },
                "required": [
                    "created_at",
                    "id",
                    "numero",
                    "statut_occupation",
                    "updated_at"
                ]
            },
...
            "CageRequest": {
                "type": "object",
                "properties": {
                    "numero": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 20
                    },
                    "nom": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "superficie": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$",
                        "nullable": true
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "numero"
                ]
            },
...
            "OccupationCouple": {
                "type": "object",
                "description": "Détail occupant couple dans la visualisation volière.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "male_bague": {
                        "type": "string"
                    },
                    "femelle_bague": {
                        "type": "string"
                    }
                },
                "required": [
                    "femelle_bague",
                    "id",
                    "male_bague"
                ]
            },
            "OccupationPigeon": {
                "type": "object",
                "description": "Détail occupant pigeon dans la visualisation volière.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "bague": {
                        "type": "string"
                    },
                    "sexe": {
                        "type": "string"
                    },
                    "race": {
                        "type": "string"
                    }
                },
                "required": [
                    "bague",
                    "id",
                    "race",
                    "sexe"
                ]
            },
...
            "PaginatedVolereCageList": {
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "properties": {
                    "count": {
                        "type": "integer",
                        "example": 123
                    },
                    "next": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VolereCage"
                        }
                    }
                }
            },
...
            "StatutOccupationEnum": {
                "enum": [
                    "LIBRE",
                    "OCCUPE_PIGEON",
                    "OCCUPE_COUPLE"
                ],
                "type": "string",
                "description": "* `LIBRE` - Libre\n* `OCCUPE_PIGEON` - Occupée par un pigeon\n* `OCCUPE_COUPLE` - Occupée par un couple"
            },
...
            "VolereCage": {
                "type": "object",
                "description": "Une case de la grille de visualisation (US-CAG-01).",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "numero": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "statut_occupation": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "pigeon": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/OccupationPigeon"
                            }
                        ],
                        "nullable": true
                    },
                    "couple": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/OccupationCouple"
                            }
                        ],
                        "nullable": true
                    }
                },
                "required": [
                    "color",
                    "couple",
                    "id",
                    "nom",
                    "numero",
                    "pigeon",
                    "statut_occupation"
                ]
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT access token obtenu via /api/v1/auth/login/"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}